+++ internal/infra/push/datadestination_oracle.go
@@ -182,7 +182,7 @@ func (d OracleDialect) IsDuplicateError(err error) bool {
 
 // ConvertValue before load
 func (d OracleDialect) ConvertValue(from push.Value, descriptor ValueDescriptor) push.Value {
-	if descriptor.column.Import() == "file" || descriptor.column.Import() == "blob" {
+	if descriptor.column != nil && (descriptor.column.Import() == "file" || descriptor.column.Import() == "blob") {
 		//nolint:gocritic
 		switch v := from.(type) {
 		case []byte: